refactor: 表記揺れ修正
